feat: fullch strip GUI - split Eq part to own component.
