+++ client/components/ChanStripFull.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import ReactSlider from 'react-slider'
-import Draggable from 'react-draggable'
 
 import '../assets/css/ChanStripFull.css'
 import { Store } from 'redux'
@@ -22,6 +21,7 @@ import ClassNames from 'classnames'
 import { fxParamsList } from '../../server/constants/MixerProtocolInterface'
 import { IChannel } from '../../server/reducers/channelsReducer'
 import { getFaderLabel } from '../utils/labels'
+import ChanStripEq from './ChanStripEq'
 
 interface IChanStripFullInjectProps {
     label: string
@@ -36,61 +36,6 @@ interface IChanStripFullInjectProps {
 interface IChanStripFullProps {
     faderIndex: number
 }
-
-enum EqColors {
-    'rgb(93, 184, 180)',
-    'rgb(53, 112, 127)',
-    'rgb(217, 21, 133)',
-    'rgb(229, 159, 34)',
-}
-
-interface IFreqLabels {
-    label: string
-    posY: number
-}
-const EQ_FREQ_LABELS: IFreqLabels[] = [
-    {
-        label: '50',
-        posY: 400,
-    },
-    {
-        label: '100',
-        posY: 525,
-    },
-    {
-        label: '250',
-        posY: 700,
-    },
-    {
-        label: '500',
-        posY: 850,
-    },
-    {
-        label: '1k',
-        posY: 1025,
-    },
-    {
-        label: '2k',
-        posY: 1160,
-    },
-    {
-        label: '5k',
-        posY: 1350,
-    },
-    {
-        label: '10k',
-        posY: 1490,
-    },
-]
-
-// Constant for calculation Eq dot positions:
-const EQ_MIN_HZ = 20
-const EQ_MAX_HZ = 20000
-const EQ_X_SIZE = 1450
-const EQ_WIN_X = 450
-const EQ_X_OFFSET = 350
-const EQ_Y_SIZE = 330
-const EQ_Y_OFFSET = 840
 
 // Constants for Delay buttons:
 const DEL_VALUES = [10, 1, -1, -10]
@@ -171,230 +116,6 @@ class ChanStripFull extends React.PureComponent<
         })
     }
 
-    handleDragCaptureEq(key: number, event: any) {
-        let eqFreqKey =
-            fxParamsList[
-                String(fxParamsList[key]).replace(
-                    'EqGain',
-                    'EqFreq'
-                ) as keyof typeof fxParamsList
-            ]
-        let eventX = event.clientX ?? event.touches[0].clientX
-        let eventY = event.clientY ?? event.touches[0].clientY
-        this.handleFx(eqFreqKey, this.freqPositionToValue(eventX))
-        this.handleFx(
-            key,
-            Math.round((100 * (EQ_Y_OFFSET - eventY)) / EQ_Y_SIZE) / 100
-        )
-    }
-
-    valueToFreqPosition(value: number) {
-        return EQ_X_SIZE * value
-    }
-    freqPositionToValue(position: number) {
-        return (position - EQ_X_OFFSET) / EQ_X_SIZE
-    }
-
-    logOscToLinFreq(value: number) {
-        return Math.round(
-            Math.pow(
-                10,
-                value * (Math.log10(EQ_MAX_HZ) - Math.log10(EQ_MIN_HZ)) +
-                    Math.log10(EQ_MIN_HZ)
-            )
-        )
-    }
-
-    setRef = (el: HTMLCanvasElement) => {
-        this.canvas = el
-        this.paintEqBackground()
-    }
-
-    paintEqBackground() {
-        if (!this.canvas) {
-            return
-        }
-        this.canvas.width = this.canvas.clientWidth
-        this.canvas.height = this.canvas.clientHeight
-        const context = this.canvas.getContext('2d', {
-            antialias: false,
-            stencil: false,
-            preserveDrawingBuffer: true,
-        }) as CanvasRenderingContext2D
-
-        if (!context) return
-
-        // Draw X-Y axis:
-        context.beginPath()
-        context.strokeStyle = 'white'
-        context.moveTo(55, 0)
-        context.lineTo(55, this.canvas.height)
-        context.stroke()
-        // Draw zero gain line:
-        context.beginPath()
-        context.strokeStyle = 'rgba(128, 128, 128, 0.244) 10px'
-        context.moveTo(55, this.canvas.height/2)
-        context.lineTo(this.canvas.width, this.canvas.height/2)
-        context.stroke()
-        // Freq on zero gain line:
-        context.beginPath()
-        EQ_FREQ_LABELS.forEach((freq: IFreqLabels) => {
-            context.font = String(this.canvas.height/ 20) + 'px Ariel'
-            context.strokeStyle = 'white'
-            context.strokeText(freq.label, freq.posY, this.canvas.height/2 + 20)
-        })
-        // Freq on zero gain line:
-        context.strokeText(
-            String(
-                window.mixerProtocol.channelTypes[0].fromMixer[
-                    fxParamsList.EqGain01
-                ]?.[0].maxLabel
-            ) + ' dB',
-            1,
-            20
-        )
-        context.strokeText('0 dB', 1, this.canvas.height/2 + 20)
-        context.strokeText(
-            String(
-                window.mixerProtocol.channelTypes[0].fromMixer[
-                    fxParamsList.EqGain01
-                ]?.[0].maxLabel
-            ) + ' dB',
-            1,
-            this.canvas.height
-        )
-        context.stroke()
-    }
-
-    eq() {
-        return (
-            <div className="eq-full">
-                <canvas className="eq-canvas" ref={this.setRef}></canvas>
-                <div className="title">EQUALIZER</div>
-                {this.eqGraphics()}
-                {this.eqText()}
-            </div>
-        )
-    }
-    eqGraphics() {
-        return (
-            <div className="eq-window">
-                {Object.keys(fxParamsList)
-                    .filter((fxKey: number | string) => {
-                        return String(fxKey).includes('EqGain')
-                    })
-                    .map((keyName: string) => {
-                        let fxKey = keyName as keyof typeof fxParamsList
-                        let eqFreqKey =
-                            fxParamsList[
-                                fxKey.replace(
-                                    'EqGain',
-                                    'EqFreq'
-                                ) as keyof typeof fxParamsList
-                            ]
-                        return (
-                            <Draggable
-                                position={{
-                                    x: this.valueToFreqPosition(
-                                        this.props.fader[this.props.faderIndex][
-                                            eqFreqKey
-                                        ]?.[0]
-                                    ),
-                                    y:
-                                        EQ_Y_SIZE -
-                                        this.props.fader[this.props.faderIndex][
-                                            fxParamsList[fxKey]
-                                        ]?.[0] *
-                                            EQ_Y_SIZE,
-                                }}
-                                grid={[20, 20]}
-                                scale={100}
-                                onDrag={(event) =>
-                                    this.handleDragCaptureEq(
-                                        fxParamsList[fxKey],
-                                        event
-                                    )
-                                }
-                            >
-                                <div
-                                    className="dot"
-                                    style={{
-                                        color: String(
-                                            EqColors[fxParamsList[fxKey]]
-                                        ),
-                                    }}
-                                >
-                                    O
-                                </div>
-                            </Draggable>
-                        )
-                    })}
-            </div>
-        )
-    }
-
-    eqText() {
-        return (
-            <div className="eq-text">
-                {Object.keys(fxParamsList)
-                    .filter((fxKey: number | string) => {
-                        return String(fxKey).includes('EqGain')
-                    })
-                    .map((keyName: string) => {
-                        let fxKey = keyName as keyof typeof fxParamsList
-                        let eqFreqKey =
-                            fxParamsList[
-                                fxKey.replace(
-                                    'EqGain',
-                                    'EqFreq'
-                                ) as keyof typeof fxParamsList
-                            ]
-                        let eqQKey =
-                            fxParamsList[
-                                fxKey.replace(
-                                    'EqGain',
-                                    'EqQ'
-                                ) as keyof typeof fxParamsList
-                            ]
-                        let maxGain: number =
-                            window.mixerProtocol.channelTypes[0].fromMixer[
-                                fxParamsList[fxKey]
-                            ]?.[0].maxLabel ?? 1
-                        let minGain =
-                            window.mixerProtocol.channelTypes[0].fromMixer[
-                                fxParamsList[fxKey]
-                            ]?.[0].minLabel ?? 0
-
-                        return (
-                            <div
-                                style={{
-                                    color: EqColors[fxParamsList[fxKey]],
-                                }}
-                            >
-                                <br />
-                                {'  Gain : '}
-                                {Math.round(
-                                    10 *
-                                        ((maxGain - minGain) *
-                                            (this.props.fader[
-                                                this.props.faderIndex
-                                            ][fxParamsList[fxKey]]?.[0] ?? 0) +
-                                            minGain)
-                                ) / 10}
-                                {'  Freq :'}
-                                {this.logOscToLinFreq(
-                                    this.props.fader[this.props.faderIndex][
-                                        eqFreqKey
-                                    ]?.[0] ?? 0
-                                )}
-                                {this.qFader(eqQKey)}
-                            </div>
-                        )
-                    })}
-            </div>
-        )
-    }
-
     inputSelectorButton(index: number) {
         const isActive =
             this.props.fader[this.props.faderIndex].inputSelector === index + 1
@@ -449,15 +170,15 @@ class ChanStripFull extends React.PureComponent<
             window.mixerProtocol.channelTypes[0].fromMixer
                 .CHANNEL_INPUT_GAIN?.[0].minLabel ?? 0
         return (
-            <div className="parameter-text">
+            <div className="chstrip-full-parameter-text">
                 Gain
-                <div className="parameter-mini-text">{maxLabel + ' dB'}</div>
+                <div className="chstrip-full-mini-text">{maxLabel + ' dB'}</div>
                 {window.mixerProtocol.channelTypes[0].toMixer
                     .CHANNEL_INPUT_GAIN ? (
                     <React.Fragment>
                         <ReactSlider
-                            className="chan-strip-full-fader"
-                            thumbClassName="chan-strip-full-thumb"
+                            className="chstrip-full-fader"
+                            thumbClassName="chstrip-full-thumb"
                             orientation="vertical"
                             invert
                             min={0}
@@ -473,14 +194,14 @@ class ChanStripFull extends React.PureComponent<
                         />
                     </React.Fragment>
                 ) : null}
-                <div className="parameter-mini-text">{minLabel + ' dB'}</div>
+                <div className="chstrip-full-mini-text">{minLabel + ' dB'}</div>
             </div>
         )
     }
 
     gainReduction() {
         return (
-            <div className="parameter-text">
+            <div className="chstrip-full-parameter-text">
                 Redution
                 <ReductionMeter faderIndex={this.props.faderIndex} />
             </div>
@@ -490,7 +211,7 @@ class ChanStripFull extends React.PureComponent<
         return (
             <React.Fragment>
                 {this.fxParamFader(fxParamsList.DelayTime)}
-                <div className="delayButtons">
+                <div className="chstrip-full-delay-buttons">
                     {DEL_VALUES.map((value: number) => {
                         return (
                             <button
@@ -513,67 +234,6 @@ class ChanStripFull extends React.PureComponent<
             </React.Fragment>
         )
     }
-    qFader(fxParam: fxParamsList) {
-        let maxLabel: number =
-            window.mixerProtocol.channelTypes[0].fromMixer[fxParam]?.[0]
-                .maxLabel ?? 1
-        let minLabel =
-            window.mixerProtocol.channelTypes[0].fromMixer[fxParam]?.[0]
-                .minLabel ?? 0
-        let valueLabel =
-            window.mixerProtocol.channelTypes[0].fromMixer[fxParam]?.[0]
-                .valueLabel ?? ''
-        return (
-            <div className="parameter-text">
-                <div className="parameter-mini-text">
-                    {Math.round(
-                        ((maxLabel - minLabel) *
-                            (1 -
-                                Math.pow(
-                                    1 -
-                                        this.props.fader[this.props.faderIndex][
-                                            fxParam
-                                        ]?.[0],
-                                    3
-                                )) +
-                            minLabel) *
-                            10
-                    ) / 10}
-                    {valueLabel}
-                </div>
-                <ReactSlider
-                    className="chan-strip-q"
-                    thumbClassName="chan-strip-q-thumb"
-                    orientation="horisontal"
-                    min={0}
-                    max={1}
-                    invert="true"
-                    step={0.01}
-                    value={
-                        this.props.fader[this.props.faderIndex][fxParam]?.[0]
-                    }
-                    renderThumb={(props: any, state: any) => (
-                        <div {...props}>
-                            {Math.round(
-                                ((maxLabel - minLabel) *
-                                    (1 -
-                                        Math.pow(
-                                            1 - parseFloat(state.valueNow),
-                                            3
-                                        )) +
-                                    minLabel) *
-                                    10
-                            ) / 10}
-                            {valueLabel}
-                        </div>
-                    )}
-                    onChange={(event: any) => {
-                        this.handleFx(fxParam, event)
-                    }}
-                />
-            </div>
-        )
-    }
 
     fxParamFader(fxParam: fxParamsList) {
         if (!this.doesParamExists(fxParam)) {
@@ -592,17 +252,17 @@ class ChanStripFull extends React.PureComponent<
             window.mixerProtocol.channelTypes[0].fromMixer[fxParam]?.[0]
                 .valueAsLabels
         return (
-            <div className="parameter-text">
+            <div className="chstrip-full-parameter-text">
                 {window.mixerProtocol.channelTypes[0].fromMixer[fxParam][0]
                     .label ?? ''}
-                <div className="parameter-mini-text">
+                <div className="chstrip-full-mini-text">
                     {!valueAsLabels
                         ? maxLabel + valueLabel
                         : valueAsLabels[valueAsLabels.length - 1] + valueLabel}
                 </div>
                 <ReactSlider
-                    className="chan-strip-full-fader"
-                    thumbClassName="chan-strip-full-thumb"
+                    className="chstrip-full-fader"
+                    thumbClassName="chstrip-full-thumb"
                     orientation="vertical"
                     invert
                     min={0}
@@ -633,7 +293,7 @@ class ChanStripFull extends React.PureComponent<
                         this.handleFx(fxParam, event)
                     }}
                 />
-                <div className="parameter-mini-text">
+                <div className="chstrip-full-mini-text">
                     {!valueAsLabels
                         ? minLabel + valueLabel
                         : valueAsLabels[0] + valueLabel}
@@ -648,7 +308,7 @@ class ChanStripFull extends React.PureComponent<
         }
         let value = this.props.fader[this.props.faderIndex][fxParam]?.[0]
         return (
-            <div className="parameter-text">
+            <div className="chstrip-full-parameter-text">
                 <div className="parameter-button-text">
                     {window.mixerProtocol.channelTypes[0].fromMixer[fxParam][0]
                         .label ?? ''}
@@ -679,9 +339,10 @@ class ChanStripFull extends React.PureComponent<
         return (
             <li key={channelIndex}>
                 {monitorName}
+                <div className="chstrip-full-mini-text">&nbsp;</div>
                 <ReactSlider
-                    className="chan-strip-full-fader"
-                    thumbClassName="chan-strip-full-thumb"
+                    className="chstrip-full-fader"
+                    thumbClassName="chstrip-full-thumb"
                     orientation="vertical"
                     invert
                     min={0}
@@ -696,7 +357,7 @@ class ChanStripFull extends React.PureComponent<
                         this.handleMonitorLevel(event, channelIndex)
                     }}
                 />
-                <p className="zero-monitor">_______</p>
+                <p className="chstrip-full-zero-monitor">_______</p>
             </li>
         )
     }
@@ -713,13 +374,13 @@ class ChanStripFull extends React.PureComponent<
                 window.mixerProtocol.channelTypes[0].toMixer
                     .CHANNEL_INPUT_SELECTOR
             return (
-                <div className="parameters">
+                <div className="chstrip-full-parameters">
                     <div className="horizontal">
                         {hasInput && (
                             <React.Fragment>
-                                <div className="item">
+                                <div className="chstrip-full-content">
                                     <div className="title">INPUT</div>
-                                    <div className="content">
+                                    <div className="chstrip-full-content">
                                         {this.inputSelector()}
                                         {this.inputGain()}
                                     </div>
@@ -728,9 +389,9 @@ class ChanStripFull extends React.PureComponent<
                         )}
                         <React.Fragment>
                             {this.doesParamExists(fxParamsList.GainTrim) ? (
-                                <div className="item">
+                                <div className="chstrip-full-content">
                                     <div className="title">INPUT</div>
-                                    <div className="content">
+                                    <div className="chstrip-full-content">
                                         {this.fxParamFader(
                                             fxParamsList.GainTrim
                                         )}
@@ -740,20 +401,24 @@ class ChanStripFull extends React.PureComponent<
                                 <React.Fragment></React.Fragment>
                             )}
                             {this.doesParamExists(fxParamsList.CompThrs) ? (
-                                <div className="item">
+                                <div className="chstrip-full-content">
                                     <div className="title">COMPRESSOR</div>
-                                    <div className="content">
+                                    <div className="chstrip-full-content">
                                         {this.fxParamButton(
                                             fxParamsList.CompOnOff
                                         )}
                                         {this.fxParamFader(
                                             fxParamsList.CompThrs
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                         {this.fxParamFader(
                                             fxParamsList.CompRatio
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                         {this.gainReduction()}
                                         <p className="reduction-6db">___6dB</p>
                                         <p className="reduction-12db">
@@ -762,41 +427,49 @@ class ChanStripFull extends React.PureComponent<
                                         {this.fxParamFader(
                                             fxParamsList.CompMakeUp
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                         {this.fxParamFader(
                                             fxParamsList.CompAttack
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                         {this.fxParamFader(
                                             fxParamsList.CompHold
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                         {this.fxParamFader(
                                             fxParamsList.CompRelease
                                         )}
-                                        <p className="zero-comp">______</p>
+                                        <p className="chstrip-full-zero-comp">
+                                            ______
+                                        </p>
                                     </div>
                                 </div>
                             ) : (
                                 <React.Fragment></React.Fragment>
                             )}
                             {this.doesParamExists(fxParamsList.DelayTime) ? (
-                                <div className="item">
+                                <div className="chstrip-full-content">
                                     <div className="title">DELAY</div>
-                                    <div className="content">
+                                    <div className="chstrip-full-content">
                                         {this.delay()}
                                     </div>
                                 </div>
                             ) : (
                                 <div className="noDelayButtons"></div>
                             )}
-                            <div className="item">
+                            <div className="chstrip-full-content">
                                 <div className="title">
                                     {this.props.label}
                                     {' - MONITOR MIX MINUS'}
                                 </div>
-                                <div className="content">
-                                    <ul className="monitor-sends">
+                                <div className="chstrip-full-content">
+                                    <ul className="chstrip-full-monitor-sends">
                                         {this.props.channel.map(
                                             (ch: any, index: number) => {
                                                 if (
@@ -816,7 +489,11 @@ class ChanStripFull extends React.PureComponent<
                     <React.Fragment>
                         <hr />
                         {this.doesParamExists(fxParamsList.EqGain01) ? (
-                            <div className="horizontal">{this.eq()}</div>
+                            <div className="horizontal">
+                                <ChanStripEq
+                                    faderIndex={this.props.faderIndex}
+                                />
+                            </div>
                         ) : (
                             <React.Fragment></React.Fragment>
                         )}
@@ -831,7 +508,7 @@ class ChanStripFull extends React.PureComponent<
     render() {
         if (this.props.faderIndex >= 0) {
             return (
-                <div className="chan-strip-full-body">
+                <div className="chstrip-full-body">
                     <div className="ch-strip-full-header">
                         {this.props.label}
                         <button
@@ -862,7 +539,7 @@ class ChanStripFull extends React.PureComponent<
                 </div>
             )
         } else {
-            return <div className="chan-strip-full-body"></div>
+            return <div className="chstrip-full-body"></div>
         }
     }
 }
